requirements.txt
requests
supabase
pandas
streamlit
plotly
folium
streamlit-folium
python-dotenv